mi: make management interface support optional
The MI code depends on Linux-specific interfaces (networking), which
makes porting to other platforms more difficult. Make it optional. This
also helps reduce the footprint for minimal setups.
